Make login a named route
https://youtu.be/M1HMtm6hj5Q?si=abzIHUVlSkg2MJvs&t=858

+++ app/Http/Controllers/JobController.php
@@ -67,6 +67,8 @@ class JobController extends Controller
 
     public function update(Job $job)
     {
+        Gate::authorize('edit-job', $job);
+
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required'],
@@ -83,7 +85,9 @@ class JobController extends Controller
 
     public function destroy(Job $job)
     {
-        // $job->delete();
+
+        Gate::authorize('edit-job', $job);
+
         $job->delete();
         return redirect('/jobs');
     }
